[TWINFACES-282] refactoring fields table
