--- src/entities/fields/api/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from "./useSearchFields";
