--- src/entities/fields/api/index.ts
@@ -1,3 +1,0 @@
-export * from "./api-service";
-export * from "./hooks";
-export * from "./types";
