+++ src/entities/twinClassField/api/index.ts
@@ -1,2 +1,3 @@
 export * from "./api-service";
 export * from "./types";
+export * from "./hooks";
